subscriber: impl Subscribe for FilterFn and DynFilterFn (#1546)
## Motivation

Currently, the `FilterFn` and `DynFilterFn` filters are only usable as
`Filter`s for per-subscriber filtering. However, there's no real reason
they can't _also_ have `Subscribe` implementations, making them usable
as global filters as well.

## Solution

This branch adds `Subscribe` implementations to `FilterFn` and
`DynFilterFn`, and moves them into their own file. It also changes the
feature flagging so that the "registry" feature is only required for the
`Filter` implementations --- the types themselves don't involve any
feature-flagged code, only the `Filter` trait.
